Authortime shared assets merged into templates
Assets.FLA has all shared assets. Depending on the template. The assets
will be pulled from other Shared files as well (AssetsDuck.fla,
AssetsFalcon.fla, Etc.).  The purpose of this is to ensure that all
assets are universal in likeness, naming and can be easily changed
throughout all templates.   How it works: If an assets has changed in
the Assets.FLA, and you publish any template sharing that library. Upon
publish, Flash will check and update with the latest assets. This will
allow for easier updates and consistency. It will not however reduce
file size since it is authortime only and the Template FLA's keep a copy
as well. It just updates dynamically.